Use a manual registration process to register built-in NodeDrivers. Can go back to a more sane module hierarchy. Closes #264.
