+++ src/feditest/cli/__init__.py
@@ -61,10 +61,5 @@ def find_commands() -> dict[str,ModuleType]:
     return cmds
 
 
-default_node_drivers_dir = [ feditest.__file__[ : -11] + 'nodedrivers' ]
-"""
-Default directory for where to look for node drivers
-"""
-
 if __name__ == '__main__':
     main()
