+++ src/feditest/nodedrivers/manual/__init__.py
@@ -2,11 +2,45 @@
 A NodeDriver that supports all protocols but doesn't automate anything.
 """
 
-from feditest import nodedriver
-from feditest.nodedrivers import AbstractManualWebServerNodeDriver
+from typing import Any
+
+from feditest.protocols import Node, NodeDriver
+from feditest.protocols.fediverse import FediverseNode
+from feditest.utils import appname_validate, hostname_validate
 
 
-@nodedriver
+class AbstractManualWebServerNodeDriver(NodeDriver):
+    """
+    Abstract superclass of NodeDrivers that support all web server-side protocols but don't
+    automate anything.
+    """
+    # Python 3.12 @override
+    def _fill_in_parameters(self, rolename: str, parameters: dict[str,Any]):
+        super()._fill_in_parameters(rolename, parameters)
+        hostname = parameters.get('hostname')
+        if hostname:
+            self.prompt_user(f'Manually provision a Node for constellation role "{ rolename }"'
+                             + f' with hostname "{ hostname }" and hit return when done.')
+        else:
+            parameters['hostname'] = self.prompt_user(f'Manually provision a Node for constellation role "{ rolename }"'
+                                        + ' and enter the hostname when done: ',
+                                        parse_validate=hostname_validate)
+        app = parameters.get('app')
+        if not app:
+            parameters['app'] = self.prompt_user('Enter the name of the app you just provisioned'
+                                                 + f' at hostname { parameters["hostname"] }: ',
+                                                 parse_validate=appname_validate)
+
+
+    # Python 3.12 @override
+    def _provision_node(self, rolename: str, parameters: dict[str,Any]) -> FediverseNode:
+        return FediverseNode(rolename, parameters, self)
+
+
+    def _unprovision_node(self, node: Node) -> None:
+        self.prompt_user(f'Manually unprovision the Node for constellation role { node.rolename() } and hit return when done.')
+
+
 class ManualFediverseNodeDriver(AbstractManualWebServerNodeDriver):
     """
     A NodeDriver that supports all web server-side protocols but doesn't automate anything.
