+++ src/feditest/nodedrivers/mastodon/ubos.py
@@ -3,11 +3,10 @@
 
 from typing import Any
 
-from feditest import nodedriver
 from feditest.nodedrivers.mastodon import MastodonNode
 from feditest.ubos import UbosNodeDriver
 
-@nodedriver
+
 class MastodonUbosNodeDriver(UbosNodeDriver):
     """
     Knows how to instantiate Mastodon via UBOS.
