+++ src/feditest/nodedrivers/saas/__init__.py
@@ -5,13 +5,11 @@ Node under test exists as a website that we don't have/can provision/unprovision
 
 from typing import Any
 
-from feditest import nodedriver
 from feditest.protocols import NodeDriver
 from feditest.protocols.fediverse import FediverseNode
 from feditest.utils import appname_validate, hostname_validate
 
 
-@nodedriver
 class SaasFediverseNodeDriver(NodeDriver):
     """
     A NodeDriver that supports all protocols but doesn't automate anything and assumes the
